EF mappings

+++ src/TimeGo.Web.Mvc/Services/IAuthorizationService.cs
@@ -1,6 +1,6 @@
 ﻿using System.Collections.Generic;
 using System.Security.Claims;
-using TimeGo.Data;
+using TimeGo.ApplicationDomain.Entities;
 using TimeGo.Web.Mvc.Models;
 
 namespace TimeGo.Web.Mvc.Services
@@ -8,6 +8,6 @@ namespace TimeGo.Web.Mvc.Services
     public interface IAuthorizationService
     {
         Employee GetUser(IEnumerable<Claim> claims);
-        AuthorizationResponse Authorization(string login, string password, int companyId);
+        AuthorizationResponse Authorization(string login, string password, long companyId);
     }
 }
